+++ mne/viz/tests/test_3d.py
@@ -19,7 +19,8 @@ from mne import (make_field_map, pick_channels_evoked, read_evokeds,
 from mne.io import read_raw_ctf, read_raw_bti, read_raw_kit, read_info
 from mne.io.meas_info import write_dig
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
-                     plot_trans, snapshot_brain_montage, plot_head_positions)
+                     plot_trans, snapshot_brain_montage, plot_head_positions,
+                     plot_alignment)
 from mne.viz.utils import _fake_click
 from mne.utils import (requires_mayavi, requires_pysurfer, run_tests_if_main,
                        _import_mlab, _TempDir, requires_nibabel, check_version)
@@ -124,7 +125,7 @@ def test_plot_evoked_field():
 
 @testing.requires_testing_data
 @requires_mayavi
-def test_plot_trans():
+def test_plot_alignment():
     """Test plotting of -trans.fif files and MEG sensor layouts."""
     # generate fiducials file for testing
     tempdir = _TempDir()
@@ -151,40 +152,43 @@ def test_plot_trans():
     )
     for system, info in infos.items():
         ref_meg = False if system == 'KIT' else True
-        plot_trans(info, trans_fname, subject='sample', meg_sensors=True,
-                   subjects_dir=subjects_dir, ref_meg=ref_meg)
+        plot_alignment(info, trans_fname, subject='sample', meg_sensors=True,
+                       subjects_dir=subjects_dir, ref_meg=ref_meg)
         mlab.close(all=True)
     # KIT ref sensor coil def is defined
-    plot_trans(infos['KIT'], None, meg_sensors=True, ref_meg=True)
+    plot_alignment(infos['KIT'], None, meg_sensors=True, ref_meg=True)
     mlab.close(all=True)
     info = infos['Neuromag']
-    assert_raises(TypeError, plot_trans, 'foo', trans_fname,
+    assert_raises(TypeError, plot_alignment, 'foo', trans_fname,
                   subject='sample', subjects_dir=subjects_dir)
-    assert_raises(TypeError, plot_trans, info, trans_fname,
+    assert_raises(TypeError, plot_alignment, info, trans_fname,
                   subject='sample', subjects_dir=subjects_dir, src='foo')
-    assert_raises(ValueError, plot_trans, info, trans_fname,
+    assert_raises(ValueError, plot_alignment, info, trans_fname,
                   subject='fsaverage', subjects_dir=subjects_dir,
                   src=sample_src)
     sample_src.plot(subjects_dir=subjects_dir)
     mlab.close(all=True)
     # no-head version
-    plot_trans(info, None, meg_sensors=True, dig=True, coord_frame='head')
+    plot_alignment(info, None, meg_sensors=True, dig=True, coord_frame='head')
     mlab.close(all=True)
     # all coord frames
     for coord_frame in ('meg', 'head', 'mri'):
-        plot_trans(info, meg_sensors=True, dig=True, coord_frame=coord_frame,
-                   trans=trans_fname, subject='sample',
-                   mri_fiducials=fiducials_path, subjects_dir=subjects_dir)
+        plot_alignment(info, meg_sensors=True, dig=True,
+                       coord_frame=coord_frame, trans=trans_fname,
+                       subject='sample', mri_fiducials=fiducials_path,
+                       subjects_dir=subjects_dir)
         mlab.close(all=True)
     # EEG only with strange options
     evoked_eeg_ecog = evoked.copy().pick_types(meg=False, eeg=True)
     evoked_eeg_ecog.info['projs'] = []  # "remove" avg proj
     evoked_eeg_ecog.set_channel_types({'EEG 001': 'ecog'})
     with warnings.catch_warnings(record=True) as w:
-        plot_trans(evoked_eeg_ecog.info, subject='sample', trans=trans_fname,
-                   source='outer_skin', meg_sensors=True, skull=True,
-                   eeg_sensors=['original', 'projected'], ecog_sensors=True,
-                   brain='white', head=True, subjects_dir=subjects_dir)
+        plot_alignment(evoked_eeg_ecog.info, subject='sample',
+                       trans=trans_fname, source='outer_skin',
+                       meg_sensors=True, skull=True,
+                       eeg_sensors=['original', 'projected'],
+                       ecog_sensors=True, brain='white', head=True,
+                       subjects_dir=subjects_dir)
     mlab.close(all=True)
     assert_true(['Cannot plot MEG' in str(ww.message) for ww in w])
 
@@ -193,20 +197,20 @@ def test_plot_trans():
                                         'sample-1280-1280-1280-bem-sol.fif'))
     bem_surfs = read_bem_surfaces(op.join(subjects_dir, 'sample', 'bem',
                                           'sample-1280-1280-1280-bem.fif'))
-    plot_trans(info, trans_fname, subject='sample', meg_sensors=True,
-               subjects_dir=subjects_dir, head=True, brain=True,
-               eeg_sensors='projected', skull=['inner_skull', 'outer_skull'],
-               source=None, bem=sphere)
-    plot_trans(info, trans_fname, subject='sample', meg_sensors=False,
-               subjects_dir=subjects_dir, head=True, brain='inflated',
-               skull=True, source=None, bem=bem_sol)
-    plot_trans(info, trans_fname, subject='sample', meg_sensors=True,
-               subjects_dir=subjects_dir, head=True, brain=False,
-               skull=True, source=None, bem=bem_surfs)
+    plot_alignment(info, trans_fname, subject='sample', meg_sensors=True,
+                   subjects_dir=subjects_dir, head=True, brain=True,
+                   eeg_sensors='projected', bem=sphere,
+                   skull=['inner_skull', 'outer_skull'], source=None)
+    plot_alignment(info, trans_fname, subject='sample', meg_sensors=False,
+                   subjects_dir=subjects_dir, head=True, brain='inflated',
+                   skull=True, source=None, bem=bem_sol)
+    plot_alignment(info, trans_fname, subject='sample', meg_sensors=True,
+                   subjects_dir=subjects_dir, head=True, brain=False,
+                   skull=True, source=None, bem=bem_surfs)
     sphere = make_sphere_model('auto', None, evoked.info)  # one layer
-    plot_trans(info, trans_fname, subject='sample', meg_sensors='helmet',
-               subjects_dir=subjects_dir, head=False, brain=True,
-               skull=False, source=None, bem=sphere)
+    plot_alignment(info, trans_fname, subject='sample', meg_sensors='helmet',
+                   subjects_dir=subjects_dir, head=False, brain=True,
+                   skull=False, source=None, bem=sphere)
 
 
 @testing.requires_testing_data
@@ -322,7 +326,7 @@ def test_snapshot_brain_montage():
     """Test snapshot brain montage."""
     info = read_info(evoked_fname)
     fig = plot_trans(info, trans=None, subject='sample',
-                     subjects_dir=subjects_dir)
+                     subjects_dir=subjects_dir)  # deprecated, for coverage
 
     xyz = np.vstack([ich['loc'][:3] for ich in info['chs']])
     ch_names = [ich['ch_name'] for ich in info['chs']]